+++ src/onegov.ballot/onegov/ballot/models/election/election.py
@@ -162,8 +162,8 @@ class Election(Base, ContentMixin, TimestampMixin,
         order_by='ElectionResult.district, ElectionResult.name',
     )
 
-    #: The total elegible voters
-    elegible_voters = summarized_property('elegible_voters')
+    #: The total eligible voters
+    eligible_voters = summarized_property('eligible_voters')
 
     #: The total recceived ballots
     received_ballots = summarized_property('received_ballots')
@@ -305,7 +305,7 @@ class Election(Base, ContentMixin, TimestampMixin,
             ElectionResult.name,
             ElectionResult.entity_id,
             ElectionResult.counted,
-            ElectionResult.elegible_voters,
+            ElectionResult.eligible_voters,
             ElectionResult.received_ballots,
             ElectionResult.blank_ballots,
             ElectionResult.invalid_ballots,
@@ -347,7 +347,7 @@ class Election(Base, ContentMixin, TimestampMixin,
             row['entity_name'] = result[9]
             row['entity_id'] = result[10]
             row['entity_counted'] = result[11]
-            row['entity_elegible_voters'] = result[12]
+            row['entity_eligible_voters'] = result[12]
             row['entity_received_ballots'] = result[13]
             row['entity_blank_ballots'] = result[14]
             row['entity_invalid_ballots'] = result[15]