+++ src/onegov.ballot/onegov/ballot/models/vote/ballot.py
@@ -123,8 +123,8 @@ class Ballot(Base, TimestampMixin, TitleTranslationsMixin,
     #: the total invalid votes
     invalid = summarized_property('invalid')
 
-    #: the total elegible voters
-    elegible_voters = summarized_property('elegible_voters')
+    #: the total eligible voters
+    eligible_voters = summarized_property('eligible_voters')
 
     def aggregate_results(self, attribute):
         """ Gets the sum of the given attribute from the results. """