+++ src/onegov.ballot/onegov/ballot/upgrade.py
@@ -329,20 +329,14 @@ def add_region_domain(context):
     tmp_type.drop(context.operations.get_bind(), checkfirst=False)
 
 
-@upgrade_task('Add election composites', always_run=True)
-def add_election_composites(context):
-    inspector = Inspector(context.operations_connection)
-    if 'election_composites' not in inspector.get_table_names(context.schema):
-        return False
-
-    if context.has_column('elections', 'composite_id'):
-        return False
-
-    context.operations.add_column(
-        'elections', Column(
-            'composite_id',
-            Text,
-            ForeignKey('election_composites.id'),
-            nullable=True
-        )
+@upgrade_task('Rename eligible voters columns')
+def renmame_elegible_voters_columns(context):
+    tables = (
+        'elections', 'election_results', 'ballots', 'ballot_results', 'votes'
     )
+
+    for table in tables:
+        if context.has_column(table, 'elegible_voters'):
+            context.operations.alter_column(
+                table, 'elegible_voters', new_column_name='eligible_voters'
+            )